Create LoadingIcon folder under component and move LoadingIcon
This is both for the component and it's CSS file.
