+++ src/components/LoadingIcon/LoadingIcon.js
